robo-architect/pyproject.toml
[project]
name = "msaez2"
version = "0.1.0"
description = "Ontology-driven Event Storming & Impact Analysis Engine with LangGraph"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "MSAez Team" }
]

dependencies = [
    # Neo4j Driver
    "neo4j>=5.0.0",
    
    # LangChain & LangGraph
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint>=2.0.0",
    
    # FastAPI (for future API server)
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.1.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "typer>=0.12.0",

    # Optional (used by ingestion when uploading PDFs)
    "PyMuPDF>=1.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[project.scripts]
msaez = "agent.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

